+++ src/components/FieldDynamicArray.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import InputDynamicArray, {FieldDynamicArrayProps as InputDynamicArrayProps} from './InputDynamicArray'
 import Fieldset from './Fieldset'
 
@@ -6,10 +5,12 @@ type FieldDynamicArrayProps = InputDynamicArrayProps & {
   name?: string
 };
 
-export default class FieldDynamicArray extends React.Component<FieldDynamicArrayProps> {
-  render() {
-    return <Fieldset label={this.props.label}>
-      <InputDynamicArray {...this.props} />
+const FieldDynamicArray: React.FC<FieldDynamicArrayProps> = (props) => {
+  return (
+    <Fieldset label={props.label}>
+      <InputDynamicArray {...props} />
     </Fieldset>
-  }
-}
+  );
+};
+
+export default FieldDynamicArray;