+++ src/components/FieldNumber.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import InputNumber, {InputNumberProps} from './InputNumber'
 import Block from './Block'
 
@@ -11,10 +10,12 @@ type FieldNumberProps = InputNumberProps & {
 };
 
 
-export default class FieldNumber extends React.Component<FieldNumberProps> {
-  render() {
-    return <Block label={this.props.label} fieldSpec={this.props.fieldSpec}>
-      <InputNumber {...this.props} />
+const FieldNumber: React.FC<FieldNumberProps> = (props) => {
+  return (
+    <Block label={props.label} fieldSpec={props.fieldSpec}>
+      <InputNumber {...props} />
     </Block>
-  }
-}
+  );
+};
+
+export default FieldNumber;