+++ CHANGELOG.md
@@ -11,11 +11,17 @@
 - Upgrade to MapLibre LG JS v5
 - Upgrade Vite 6 and Cypress 14 ([#970](https://github.com/maplibre/maputnik/pull/970))
 - Upgrade OpenLayers from v6 to v10
+- When loading a style into localStorage that causes a QuotaExceededError, purge localStorage and retry
+- Remove react-autobind dependency
+- Remove usage of legacy `childContextTypes` API
+- Refactor Field components to use arrow function syntax
+- Replace react-autocomplete with Downshift in the autocomplete component
 - _...Add new stuff here..._
 
 ### 🐞 Bug fixes
 
 - Fix incorrect handing of network error response (#944)
+- Show an error when adding a layer with a duplicate ID
 - _...Add new stuff here..._
 
 ## 2.1.1